I realized now that the poll detection is very flaky. Tony Hawk's Pro Skater was broken because poll detection of gpuctrl reads wasn't happening, and interlaced vsync wasn't being forceably flipped.
The read of this value is not in a tight loop, but rather in a multi-function thing mixed with (I think) root counter checks.

Anyway, stage1/2 compilers affect this because the whole deciding when to call tagRead32 and checkPoll32 etc. is poorly thought out (the effectiveness of checkPoll32 depends on how many call sites it has).

Anyway, I hacked the 32bit poll path to try to make it detect more h/w polls, but also added a fall back in the GPU to detect polling there too (with a much higher loop count)